Issue #10574:add delete confirmation
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@8151 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/vendors.cpp
@@ -109,6 +109,13 @@ void vendors::sCopy()
 
 void vendors::sDelete()
 {
+  QString question = tr("Are you sure that you want to delete this vendor?");
+  if (QMessageBox::question(this, tr("Delete Vendor?"),
+                              question,
+                              QMessageBox::Yes,
+                              QMessageBox::No | QMessageBox::Default) == QMessageBox::No)
+    return;
+
   q.prepare("SELECT deleteVendor(:vend_id) AS result;");
   q.bindValue(":vend_id", _vendor->id());
   q.exec();
